delete itlwmx, organize project structure
