--- ItlwmFirmwareGen1/ItlwmFirmwareGen1.cpp
@@ -1,4 +1,0 @@
-/* add your code here */
-
-#include "ItlwmFirmwareGen1.hpp"
-
