--- ItlwmFirmwareGen1/ItlwmFirmwareGen1.hpp
@@ -1,10 +1,0 @@
-/* add your code here */
-
-#include "FirmwareStoreService.h"
-
-class ItlwmFirmwareGen1 : FirmwareStoreService {
-    
-    
-public:
-    
-};
