--- ItlwmFirmwareGen2/Info.plist
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CFBundleDevelopmentRegion</key>
-	<string>$(DEVELOPMENT_LANGUAGE)</string>
-	<key>CFBundleExecutable</key>
-	<string>$(EXECUTABLE_NAME)</string>
-	<key>CFBundleIdentifier</key>
-	<string>$(PRODUCT_BUNDLE_IDENTIFIER)</string>
-	<key>CFBundleInfoDictionaryVersion</key>
-	<string>6.0</string>
-	<key>CFBundleName</key>
-	<string>$(PRODUCT_NAME)</string>
-	<key>CFBundlePackageType</key>
-	<string>$(PRODUCT_BUNDLE_PACKAGE_TYPE)</string>
-	<key>CFBundleShortVersionString</key>
-	<string>1.0</string>
-	<key>CFBundleVersion</key>
-	<string>1</string>
-	<key>IOKitPersonalities</key>
-	<dict>
-		<key>ItlwmFirmwareGen2</key>
-		<dict>
-			<key>CFBundleIdentifier</key>
-			<string>com.zxystd.ItlwmFirmwareGen2</string>
-			<key>IOClass</key>
-			<string>ItlwmFirmwareGen2</string>
-			<key>IOMatchCategory</key>
-			<string>ItlwmFirmwareGen2</string>
-			<key>IOProviderClass</key>
-			<string>IOResources</string>
-		</dict>
-	</dict>
-	<key>NSHumanReadableCopyright</key>
-	<string>Copyright © 2020 钟先耀. All rights reserved.</string>
-	<key>OSBundleLibraries</key>
-	<dict>
-		<key>com.apple.kpi.bsd</key>
-		<string>16.7</string>
-		<key>com.apple.kpi.iokit</key>
-		<string>16.7</string>
-		<key>com.apple.kpi.libkern</key>
-		<string>16.7</string>
-		<key>com.apple.kpi.mach</key>
-		<string>16.7</string>
-	</dict>
-</dict>
-</plist>
