--- ItlwmFirmwareGen2/ItlwmFirmwareGen2.cpp
@@ -1,1 +1,0 @@
-/* add your code here */
