--- ItlwmFirmwareGen2/ItlwmFirmwareGen2.hpp
@@ -1,9 +1,0 @@
-/* add your code here */
-#include "FirmwareStoreService.h"
-
-class ItlwmFirmwareGen2 : FirmwareStoreService {
-    
-    
-public:
-    
-};
