+++ include/ClientKit/Common.h
