+++ itlwm/firmware/iwlwifi-Qu-b0-hr-b0-48.ucode
