+++ itlwm/firmware/iwlwifi-QuZ-a0-hr-b0-48.ucode
