+++ itlwm/hal_iwx/ItlIwx.hpp
@@ -137,9 +137,9 @@
 #include <libkern/c++/OSMetaClass.h>
 #include <IOKit/IOFilterInterruptEventSource.h>
 
-#include "ItlHalService.hpp"
-#include "ItlDriverController.h"
-#include "ItlDriverInfo.h"
+#include <HAL/ItlHalService.hpp>
+#include <HAL/ItlDriverInfo.hpp>
+#include <HAL/ItlDriverController.hpp>
 
 class ItlIwx : public ItlHalService, ItlDriverInfo, ItlDriverController {
     OSDeclareDefaultStructors(ItlIwx)
